Set up pynpoint example

+++ paco/processing/fastpaco.py
@@ -167,18 +167,19 @@ class FastPACO(PACO):
             h_template = self.m_psf
         else:
             h_template = self.modelFunction(k, model_name, params)
-        h_mask = createCircularMask(h_template.shape,radius = self.m_psf_rad*scale)
-        h = np.zeros((self.m_height,self.m_width,self.m_p_size*scale**2)) # The off axis PSF at each point
+        h_mask = createCircularMask(h_template.shape,radius = int(self.m_psf_rad*scale))
+        # The off axis PSF at each point
+        h = np.zeros((self.m_height,self.m_width,int(self.m_p_size*scale**2))) 
 
         # Store for each image pixel, for each temporal frame an image
         # for patches: for each time, we need to store a column of patches
-        patch = np.zeros((self.m_nFrames,self.m_p_size*scale**2)) # 2d selection of pixels around a given point
+        patch = np.zeros((self.m_nFrames,int(self.m_p_size*scale**2))) # 2d selection of pixels around a given point
         mask =  createCircularMask((k,k),radius = self.m_psf_rad)
 
         # the mean of a temporal column of patches centered at each pixel
-        m     = np.zeros((self.m_height,self.m_width,self.m_p_size*scale**2)) 
+        m     = np.zeros((self.m_height,self.m_width,int(self.m_p_size*scale**2))) 
         # the inverse covariance matrix at each point
-        Cinv  = np.zeros((self.m_height,self.m_width,self.m_p_size*scale,self.m_p_size*scale)) 
+        Cinv  = np.zeros((self.m_height,self.m_width,int(self.m_p_size*scale),int(self.m_p_size*scale))) 
 
         # *** SERIAL ***
         # Loop over all pixels
